sim: Fix -Werror compilation issues on 64-bit Linux
This fixes two warnings that break compilation when -Werror is
enabled for the simulator on 64-bit Linux:
1. Fixes an int64_t format warning in log_shell.c.
2. Fixes an overflow warning in shell_nlip.c where an unsigned
size_t was negated.

+++ sys/log/full/src/log_shell.c
@@ -128,7 +128,7 @@ shell_log_dump_entry(struct log *log, struct log_offset *log_offset,
             console_printf("[ih=0x%02x%02x%02x%02x]", ueh->ue_imghash[0], ueh->ue_imghash[1],
                            ueh->ue_imghash[2], ueh->ue_imghash[3]);
         }
-        console_printf(" [%llu] ", ueh->ue_ts);
+        console_printf(" [%llu] ", (unsigned long long)ueh->ue_ts);
     }
 
 #if MYNEWT_VAL(LOG_SHELL_SHOW_INDEX)
